migrate to poetry, oauth.py + client.py refactoring (incomplete)
